backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
sqlalchemy==2.0.44
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
mysql-connector-python>=8.0.33
PyMySQL>=1.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0